Handle ActionWhitelist within WhitelistedCommenter
When the action of a pull request is not in the enabled_actions list a comment
would have been made in the affirmative when it should not have been.

Previously, the logic inside ActionWhitelist was inverted such that it would
return false when the pull request's action was in the list of enabled_actions
(if they were defined), which is confusing and leads to the aforementioned
bug now that affirmative commenting is an option.

Signed-off-by: Derek Richard <drichard@pivotallabs.com>